Move SoundManager sound object control into new soundProxy
psylusPlayer now acts as just a view for the track playback and does
not manipulate the sound object directly. That is instead handled by
a new soundProxy object which is able to start playback and take
appropriate actions in response to the user manipulating the UI,
depending on the current playback state.